backend/requirements.txt
# FastAPI and ASGI Server
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# Database (SQLite - 기본 포함, 별도 설치 불필요)
sqlalchemy>=2.0.36
alembic>=1.14.0

# Pydantic for data validation (Python 3.13 호환)
pydantic>=2.10.0
pydantic-settings>=2.6.0

# External API clients
google-generativeai>=0.8.0  # Gemini API
google-cloud-speech>=2.32.0  # STT (Gemini는 STT 직접 지원 안함)
google-cloud-texttospeech>=2.18.0  # TTS

# HTTP Client
httpx>=0.27.0
aiohttp>=3.10.0

# Environment variables
python-dotenv>=1.0.1

# File handling
python-jose[cryptography]>=3.3.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0

# Development
black>=24.10.0
flake8>=7.1.0
mypy>=1.11.0

